requirements.txt
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.2.1
duckduckgo_search==8.1.1
filelock==3.20.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
fsspec==2025.9.0
greenlet==3.2.4
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.2
packaging==25.0
pillow==12.0.0
primp==0.15.0
python-pptx==1.0.2
PyYAML==6.0.3
requests==2.32.5
SQLAlchemy==2.0.41
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
Wikipedia-API==0.8.1
xlsxwriter==3.2.9
gunicorn==23.0.0
moviepy==2.1.1
